Auto-completion, bug fixes.
git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@2400 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/engine/FileIO.c
@@ -1852,7 +1852,7 @@ static int
 writeTSDate( int fd, Timespec *ts)
   {
   int err=0;
-  int tmp;
+  long int tmp;
   long long longtmp;
 
   /* write 64 bits to file format */
@@ -1864,8 +1864,8 @@ writeTSDate( int fd, Timespec *ts)
   
   tmp = ts->tv_nsec;
   XACC_FLIP_INT (tmp);
-  err = write( fd, &tmp, sizeof(int) );
-  if( err != sizeof(int) )
+  err = write( fd, &tmp, sizeof(long int) );
+  if( err != sizeof(long int) )
     return -1;
   
   return err;
